Start conversion to safer money calculations Todo: forms, tests

+++ database/migrations/2020_02_24_221841_create_users_table.php
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration
             $table->id();
             $table->string('full_name');
             $table->string('username')->unique();
-            $table->float('balance');
+            $table->integer('balance');
             $table->integer('role_id');
             $table->string('password')->nullable();
             $table->rememberToken();
